requirements.txt
nltk==3.8.1
numpy
openai==1.6.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
regex==2023.12.25
sniffio==1.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
ragas
weaviate-client
langchain
datasets
python-docx==1.1.0
docx==0.2.4
langchain_experimental
langchain_openai
sentence-transformers==2.3.1
PyMuPDF==1.23.16
textract==1.6.5
pymilvus==2.3.6
langchain-pinecone==0.0.2